Add tests for syscalls
Added tests for the following syscalls:
malloc, calloc, realloc, free, open, read, write, close

The amount of syscalls to be performed can be passed as an argument
or it will be selected randomly. Min and max values of number of
syscalls to be selected randomly can be set in test/test_opprofiler.h

Other test parameters such as min and max read/write buffer size,
testfile name, etc can be set in test/test_opprofiler.h as well.

The order of syscalls is selected randomly with rand() using time(0)
as seed.

Pointers to allocated memory and opened file descriptors are stored
in arrays on stack. These arrays are allocated before tests execution
via alloca(). Each array size is set to max number of syscalls to be performed.

When malloc() or calloc() is selected, the memory is allocated and stored
in array of allocated memory pointers.

When realloc() is selected, the last allocated memory pointer is used for
realloc(). If array of allocated memory pointers is empty, realloc is
not performed and another syscall is selected instead.

Free() syscall is performed similar to realloc().

When open() is selected, a new file is created and open, the file
descriptor is added to array of opened file descriptors.

Read() and write() are performed on last open file descriptor or
skipped. Buffer size is chosen randomly.

Close() is performed on last open file descriptor or skipped. Created
file is deleted as well.

After all tests are executed, cleanup() function closes all unclosed
files and removes them and frees allocated memory which was not freed
during tests.